requirements.txt
openai>=1.42.0
tqdm>=4.66.4
numpy>=1.26.0
scikit-learn>=1.5.0
python-dotenv>=1.0.1
langchain-community>=0.3.0
pymupdf>=1.24.9
